Containers

Description

Containers: a uniform way to work with entities that may contain other entities.

Documentation

class Eq a ⇒ Container c a | c → a where

The Container class is used for types that can contain objects and can be combined with union, intersection, and difference (relative complement). Instances of Container should satisfy the following laws:

isIn == flip contains
isNotIn == flip doesNotContain
doesNotContain a == not . contains a
contains a empty == False
contains a (singleton b) == (a == b)
contains a (insert b c) == (a == b) || contains a c
contains a (union c1 c2) == contains a c1 || contains a c2
contains a (intersection c1 c2) == contains a c1 && contains a c2
intersection c c == c
difference c c == empty

Minimal complete definition

(contains | isIn), union, intersection, difference, empty, (insert | singleton)

Methods

isIn ∷ c → a → Bool

isNotIn ∷ c → a → Bool

contains ∷ a → c → Bool

doesNotContain ∷ a → c → Bool

isEmpty ∷ Eq c ⇒ c → Bool

union ∷ c → c → c

(union a b) returns a collection of elements that are in one of a or b, or both.

intersection ∷ c → c → c

(intersection a b) returns a collection of elements that are in both a and b.

difference ∷ c → c → c

(difference a b) returns a collection of elements that are in a but not in b.

symmetricDifference ∷ c → c → c

(symmetricDifference a b) returns a collection of elements that are in one of a or b, but not both.

empty ∷ c

insert ∷ a → c → c

singleton ∷ a → c

isSubsetOf ∷ Eq c ⇒ c → c → Bool

(isSubsetOf y x) tells whether x is a subset of y.

isSupersetOf ∷ Eq c ⇒ c → c → Bool

(isSupersetOf y x) tells whether x is a superset of y.

isProperSubsetOf ∷ Eq c ⇒ c → c → Bool

(isProperSubsetOf y x) tells whether x is a proper subset of y.

isProperSupersetOf ∷ Eq c ⇒ c → c → Bool

(isProperSupersetOf y x) tells whether x is a proper superset of y.

class Linearizable l where

The Linearizable class is used for types that can be traversed linearly in one direction.

Methods

choose ∷ l a → (a, l a)

Return the next element and the collection of remaining elements.

chooseOne ∷ Linearizable l ⇒ l a → a

Like choose, but discards the remaining elements.

discardOne ∷ Linearizable l ⇒ l a → l a

Like choose, but discards the next element.

class Linearizable c ⇒ Collapsible c where

The Collapsible class is used for types that can be collapsed to a single value, like a fold over a list. Any structure \(c\) that is Collapsible must necessarily be Linearizable, since:

collapse (:) [] c

performs a linearization.

Minimal complete definition

collapse | size

Methods

collapse ∷ (a → b → b) → b → c a → b

size ∷ Integral a ⇒ c b → a

fromCollapsible ∷ (Collapsible s, Container c a) ⇒ s a → c

Build a Container from the elements of a Collapsible. This can be used to cast between most types of Container. Time complexity is \(O(nci)\), where \(n\) is the number of elements in the source, \(c\) is the cost of accessing a next element of the source, and \(i\) is the cost of inserting an element into the destination.

Combining multiple Containers

unionAll ∷ (Container c a, Collapsible s) ⇒ s c → c

Combine Containers with union.

intersectAll ∷ (Container c a, Collapsible s) ⇒ s c → c

Combine Containers with intersection. An empty source yields an empty result.

Generic versions of Prelude functions and similar

anyS ∷ Collapsible s ⇒ (a → Bool) → s a → Bool

True iff some element satisfies a predicate.

allS ∷ Collapsible s ⇒ (a → Bool) → s a → Bool

True iff all elements satisfy a predicate.

tmap ∷ (Collapsible s, Container (s b1) b) ⇒ (a → b) → s a → s b1

Appy a function to each element of a Collapsible.

keep ∷ (Collapsible s, Container (s a) a) ⇒ (a → Bool) → s a → s a

Retain only those elements that satisfy a predicate.

groupBy ∷ (Eq b, Collapsible s, Container (s a) a, Container (s (s a)) (s a)) ⇒ (a → b) → s a → s (s a)

Partition a Container. For example,

groupBy (`mod` 3) [0..9] == [[0,3,6,9],[1,4,7],[2,5,8]]

Multisets

data Multiset a

A Multiset is a Set that may contain more than one instance of any given element.

multiplicity ∷ Ord a ⇒ Multiset a → a → Integer

Analogous to isIn, returning the number of occurrences of an element in a Multiset. Time complexity is \(O(\log{n})\), where \(n\) is the number of distinct elements in the Multiset.

multisetFromList ∷ Ord a ⇒ [a] → Multiset a

A specialization of fromCollapsible.

setFromMultiset ∷ Multiset a → Set a

A specialization of fromCollapsible with time complexity \(O(n)\), where \(n\) is the number of distinct elements in the source.

Set of Set with alternate ordering

The choose instance for Set will always pick the least available element. If one wants to process elements in a different order, one can simply wrap the elements in such a way that they sort in the intended order of processing. This section contains some such wrapper types.

newtype IncreasingSize x

Wrap a Collapsible type to sort in order of increasing size. For elements of the same size, treat them normally.

Constructors

IncreasingSize

Fields

getIncreasing ∷ x

newtype DecreasingSize x

Wrap a Collapsible type to sort in order of decreasing size. For elements of the same size, treat them normally.

Constructors

DecreasingSize

Fields

getDecreasing ∷ x

Miscellaneous functions

tr

Arguments

∷ (Container (s a) a, Collapsible s, Eq a)

⇒ [a]
search

→ [a]
replacement

→ s a
string

→ s a

Translate elements. All instances of elements of the search set are replaced by the corresponding elements of the replacement set in the given string. If the replacement set is smaller than the search set, it is made longer by repeating the last element.

tr "aeiou" "x" "colorless green ideas" == "cxlxrlxss grxxn xdxxs"
tr "abcdefghijklmnopqrstuvwxyz" "nopqrstuvwxyzabcdefghijklm" "cat" == "png"