ExtractSL

Description

Find forbidden substrings of an automaton.

Documentation

data ForbiddenSubstrings e

A convenience-type for declaring collections of forbidden substrings. The member types are (lists of) the raw alphabet type (not (Symbol .))

Constructors

ForbiddenSubstrings

Fields

attestedUnits ∷ Set e

Symbols that actually label transitions

forbiddenUnits ∷ Set e

Symbols of alphabet of FSA - attestedUnits

forbiddenWords ∷ Set [e]

Sequences of symbols that cannot occur as words

forbiddenInitials ∷ Set [e]

Sequences of symbols that cannot occur initially

forbiddenFrees ∷ Set [e]

Sequences of symbols that cannot occur anywhere

forbiddenFinals ∷ Set [e]

Sequences of symbols that cannot occur finally

data ForbiddenPaths n e

Constructors

ForbiddenPaths

Fields

initialPaths ∷ Set (Path n e)

Paths witnessing forbidden initial factors

freePaths ∷ Set (Path n e)

Paths witnessing forbidden free factors

finalPaths ∷ Set (Path n e)

Paths witnessing forbidden final factors

data TaggedSubstring e

Constructors

Free [e]

Initial [e]

Final [e]

Word [e]

Extracting forbidden substrings

factorsFromPaths ∷ (Ord e, Ord n) ⇒ Set (Path n e) → Set [Symbol e]

Convert Set of Paths to Set of sequences of (Symbol e)

forbiddenSubstrings ∷ (Ord e, Ord n, Enum n) ⇒ FSA n e → ForbiddenSubstrings e

Forbidden substrings of the given FSA relative to its alphabet

forbiddenPaths ∷ (Ord n, Enum n, Ord e) ⇒ FSA n e → ForbiddenPaths (Set n) e

Paths representing forbidden substrings of the given FSA.

forbiddenPathsWithBound ∷ (Ord n, Enum n, Ord e) ⇒ Integer → FSA n e → ForbiddenPaths (Set n) e

The forbidden substrings of the given FSA. Takes at most bound-1 many iterations of each cycle in the powerset graph.

Building automata

buildFSA ∷ (NFData e, Ord e) ⇒ ForbiddenSubstrings e → FSA Integer e

Create an FSA satisfying the conditions imposed by the given sets of forbidden substrings.

Determining SL

isSL ∷ (Ord e, Ord n) ⇒ FSA n e → Bool

Returns True iff the stringset represented by the given FSA is Strictly Local, that is, if it satisfies Suffix-Substition Closure for some specific factor size \(k\).

slQ ∷ (Ord e, Ord n) ⇒ FSA n e → Integer

Returns the smallest factor size for which the stringset represented by the given FSA satisfies Suffix-Substitution Closure, or 0 if there is no such \(k\).

...with precomputed PSG

isSLPSG ∷ (Ord e, Ord n) ⇒ FSA (Set n) e → Bool

isSL with precomputed PSG

slPSGQ ∷ (Ord e, Ord n) ⇒ FSA (Set n) e → Integer

slQ with precomputed PSG