FSA

Description

The purpose of this module is to define an interface to a generic, reusable impementation of finite-state automata (FSAs). The primary motivation for this is to allow for efficient analysis of stringsets in a linguistic context, although the nature of the project should allow more general use.

Documentation

data FSA n e

A finite-state automaton (FSA) is represented by a directed graph, the edges of which are labelled by formal symbols.

Constructors

FSA

Fields

alphabet ∷ Set e

transitions ∷ Set (Transition n e)

initials ∷ Set (State n)

finals ∷ Set (State n)

isDeterministic ∷ Bool

states ∷ (Ord e, Ord n) ⇒ FSA n e → Set (State n)

The collection of all states in an FSA.

isNull ∷ (Ord e, Ord n) ⇒ FSA n e → Bool

True iff the input accepts no strings.

Constructing simple automata

totalWithAlphabet ∷ (Ord e, Enum n, Ord n) ⇒ Set e → FSA n e

An automaton accepting every string over a given alphabet.

emptyWithAlphabet ∷ (Ord e, Enum n, Ord n) ⇒ Set e → FSA n e

An automaton accepting no strings over a given alphabet.

emptyLanguage ∷ (Ord e, Ord n, Enum n) ⇒ FSA n e

A specialization of emptyWithAlphabet where the alphabet is itself empty.

singletonWithAlphabet ∷ (Ord e, Enum n, Ord n) ⇒ Set e → [e] → FSA n e

An automaton that accepts only the given string, over a given alphabet.

singletonLanguage ∷ (Ord e, Enum n, Ord n) ⇒ [e] → FSA n e

An automaton that accepts only the given string, over the minimal alphabet required to represent this string.

Derived automata

kleeneClosure ∷ (Ord n, Ord e) ⇒ FSA n e → FSA (Either n Bool) e

The Kleene Closure of an automaton is the free monoid under concatenation generated by all strings in the automaton's represented stringset. The resulting automaton is nondeterministic.

powersetGraph ∷ (Ord e, Ord n) ⇒ FSA n e → FSA (Set n) e

Given an automaton \(M\) with stateset \(Q\), the powerset graph of \(M\) is an automaton with stateset in the powerset of \(Q\). From a node \(\{q_1,q_2,\ldots,q_n\}\), there is an edge labelled \(\sigma\) that leads to \(\{\delta(q_1,\sigma), \delta(q_2,\sigma), \ldots, \delta(q_n, \sigma)\}\), where \(\delta\) is the transition function of the input. The initial state is \(Q\), and the result is complete.

syntacticMonoid ∷ (Ord e, Ord n) ⇒ FSA n e → FSA ([Maybe n], [Symbol e]) e

Given an automaton \(M\) with stateset \(Q\), the syntactic monoid of \(M\) is an automaton with stateset in \((Q\rightarrow Q)\). Here these functions are represented by lists, where \(q_i\) maps to the \(i^\text{th}\) element of the list. From a node \(\langle q_1,q_2,\ldots,q_n\rangle\), there is an edge labelled \(\sigma\) that leads to \(\langle\delta(q_1,\sigma), \delta(q_2,\sigma), \ldots, \delta(q_n, \sigma)\rangle\), where \(\delta\) is the transition function of the input. The initial state is the identity function, and the result is complete.

residue ∷ (Ord n, Ord e, Enum n) ⇒ FSA n e → FSA n e → FSA n e

(residue a b) is equivalent to (difference a b). In the context of an approximation and its source, represents the strings accepted by the approximation that should not be.

coresidue ∷ (Ord n, Ord e, Enum n) ⇒ FSA n e → FSA n e → FSA n e

(coresidue a b) is equivalent to (complement (residue a b)). In the context of an approximation and its source, represents unmet constraints of the source.

Transformations

flatIntersection ∷ (Enum n, Ord n, NFData n, Ord e, NFData e) ⇒ [FSA n e] → FSA n e

Intersect all given automata, in parallel if possible. An empty intersection is undefined. In theory it should be the total language over the total alphabet, but the latter cannot be defined. Intermediate results are evaluated to normal form.

flatUnion ∷ (Enum n, Ord n, NFData n, Ord e, NFData e) ⇒ [FSA n e] → FSA n e

Union all given automata, in parallel if possible. An empty union is defined as the empty language over an empty alphabet. Intermediate results are evaluated to normal form.

reverse ∷ (Ord e, Ord n) ⇒ FSA n e → FSA n e

The reversal of an automaton accepts the reversals of all strings accepted by the original.

complement ∷ (Ord e, Ord n) ⇒ FSA n e → FSA (Set n) e

Returns an FSA accepting all and only those strings not accepted by the input.

complementDeterministic ∷ (Ord e, Ord n) ⇒ FSA n e → FSA n e

Returns the complement of a deterministic FSA. The precondition that the input is deterministic is not checked.

determinize ∷ (Ord e, Ord n) ⇒ FSA n e → FSA (Set n) e

Returns a deterministic automaton representing the same stringset as the potentially nondeterministic input.

Minimization

minimize ∷ (Ord e, Ord n) ⇒ FSA n e → FSA (Set (Set n)) e

Returns a deterministic FSA recognizing the same stringset as the input, with a minimal number of states.

minimizeDeterministic ∷ (Ord e, Ord n) ⇒ FSA n e → FSA (Set n) e

Returns a deterministic FSA recognizing the same stringset as the input, with a minimal number of states. The precondition that the input is deterministic is not checked.

normalize ∷ (Ord e, Ord n) ⇒ FSA n e → FSA Integer e

Returns a normal form of the input. An FSA is in normal form if it is minimal and deterministic, and contains neither unreachable states nor nonaccepting sinks. Node labels are irrelevant, so Int is used as a small representation.

Equivalence Classes

minimizeOver ∷ (Ord e, Ord n) ⇒ (FSA n e → Set (Set (State n))) → FSA n e → FSA (Set n) e

Returns a deterministic FSA minimized over a given relation. The precondition that the input is deterministic is not checked.

nerode ∷ (Ord e, Ord n) ⇒ FSA n e → Set (Set (State n))

Two strings \(u\) and \(v\) are equivalent iff for all strings \(w\), \(uw\) and \(vw\) lead to states in the same equivalence class.

jEquivalence ∷ (Ord e, Ord n) ⇒ FSA ([Maybe n], [Symbol e]) e → Set (Set (State ([Maybe n], [Symbol e])))

Given an automaton whose syntactic monoid is \(M\), two strings \(u\) and \(v\) are equivalent if \(MuM=MvM\)

Alphabetic Transformations

extendAlphabetTo ∷ (Ord a, Ord b) ⇒ Set b → FSA a b → FSA (Maybe Integer, Maybe a) b

Add missing symbols to the alphabet of an automaton. The result is an automaton with at least the provided alphabet that licenses exactly the same set of strings as the input.

semanticallyExtendAlphabetTo ∷ (Ord a, Ord b) ⇒ Set b → FSA a (Maybe b) → FSA a (Maybe b)

Add missing symbols to the alphabet of an automaton. As the symbol Nothing is taken to represent any symbol not currently in the alphabet, new edges are added in parallel to existing edges labelled by Nothing.

contractAlphabetTo ∷ (Ord a, Ord b) ⇒ Set b → FSA a b → FSA a b

Remove symbols from the alphabet of an automaton.

forceAlphabetTo ∷ (Ord a, Ord b) ⇒ Set b → FSA a b → FSA (Maybe Integer, Maybe a) b

Ignore the alphabet of an automaton and use a given alphabet instead.

desemantify ∷ (Ord a, Ord b) ⇒ FSA a (Maybe b) → FSA a b

Remove the semantic Nothing edges from an automaton and reflect this change in the type.

renameSymbolsBy ∷ (Ord e, Ord e1, Ord n) ⇒ (e → e1) → FSA n e → FSA n e1

Transform the edge labels of an automaton using a given function. If this function is not injective, the resulting FSA may not be deterministic even if the original was.

Transformations of State labels

renameStatesBy ∷ (Ord e, Ord n, Ord n1) ⇒ (n → n1) → FSA n e → FSA n1 e

Transform the node labels of an automaton using a given function. If this function is not injective, the resulting FSA may not be deterministic even if the original was.

renameStates ∷ (Ord e, Ord n, Ord n1, Enum n1) ⇒ FSA n e → FSA n1 e

Equivalent to renameStatesBy \(f\), where \(f\) is an arbitrary injective function.

Miscellaneous

data State n

A vertex of the graph representation of an FSA is a State, which can be labelled with any arbitrary value, so long as every vertex of a single automaton is labelled with a distinct value of the same type.

Constructors

State

Fields

nodeLabel ∷ n

data Symbol e

The label of a Transition.

Constructors

Epsilon
The edge may be taken without consuming any input.

Symbol e
The edge requires consuming this symbol.

unsymbols ∷ (Collapsible s, Eq (s e), Container (s e) e, Monoid (s e)) ⇒ s (Symbol e) → s e

Remove Epsilon from a Collapsible of Symbol and present the unwrapped results as a new Container.

data Transition n e

The edges of an FSA.

Constructors

Transition

Fields

edgeLabel ∷ Symbol e

source ∷ State n

destination ∷ State n

module Containers