LTK.Decide.SPL

Description

A language is strictly piecewise-local if whenever a word w is accepted, and a word u has the same subsequences of length up to j of substrings of length up to k (i.e. generalized j,k-subsequences), then u, too, must be accepted.

Documentation

isSPL :: (Ord n, Ord e) => FSA n e -> Bool

True iff the language is strictly piecewise-local, characterized by a set of forbidden subsequences of substrings.

Since: 1.3

isSPLM :: (Ord n, Ord e) => SynMon n e -> Bool

True iff the language is strictly piecewise-local, characterized by a set of forbidden subsequences of substrings.

Since: 1.3

isSPLs :: FiniteSemigroupRep s => OrderedSemigroup s -> Bool

True iff the semigroup satisifes \(x^{\omega}\leq x^{\omega}yx^{\omega}\).

Since: 1.3