LTK.Learn.SL

Description

This module implements a string extension learner for the SL class.

Since: 0.3

Documentation

data SLG a

A representation of an SL grammar.

Constructors

SLG

Fields

slgAlpha :: Set a

slgK :: Int

slg :: Set (Bool, [a], Bool)

fSL :: Ord a => Int -> [a] -> SLG a

Return the set of \(k\)-factors under successor in the given word. Factors are triples, where the first and last components are Booleans that indicate whether the factor is anchored at its head or tail, respectively, and the central component is the factor itself. If a word is short enough to not contain any \(k\)-factors, the entire word, appropriately anchored, is included in the set.