Porters

Description

This module provides methods to convert automata to and from various formats.

Conversions

In the following definitions, (Type t) is shorthand for (String -> t).

from ∷ Importable i ⇒ Type i → String → FSA Integer String

Create an FSA from a String treated as the given Type.

fromE ∷ Importable i ⇒ Type i → String → Either String (FSA Integer String)

Try to create an FSA from a String treated as the given Type.

to ∷ (Ord n, Ord e, Show n, Show e, Exportable x) ⇒ Type x → FSA n e → String

Create a String from an FSA, formatted appropriately for the given Type.

Formats

We use types to create a bit of magic in order to read and write automata in various formats.

newtype Dot

The GraphViz Dot format.

Constructors

Dot String

newtype Jeff

Jeff's format.

Constructors

Jeff String

newtype Pleb

Constructors

Pleb String

Miscellaneous

formatSet ∷ Show n ⇒ Set n → String

Turn a Set into a String:

>>> formatSet (fromList [1, 2, 3])
"{1, 2, 3}"

transliterate ∷ Ord n ⇒ FSA n String → FSA n String

Automata in Jeff's format use edge labels of the form “w0.s1”. This function converts the edge labels of an automaton from this form to the “L`” form that we tend to use.

transliterateString ∷ String → String

See transliterate. This function operates directly on the representation of the edge label.

untransliterate ∷ Ord n ⇒ FSA n String → FSA n String

The inverse of transliterate.

untransliterateString ∷ String → String

The inverse of transliterateString.

class Importable t

A type that can be read and turned into an FSA.

Minimal complete definition

toFSA

class Exportable t

A type that can be written from an FSA.

Minimal complete definition

fromFSA, extract