Traversals

Description

Find paths through an automaton.

Documentation

data Path n e

A path through an FSA.

Constructors

Path

Fields

labels ∷ [Symbol e]

Edge labels are gathered in reverse order, so labels is a reversed string.

endstate ∷ Maybe (State n)

Current State, if any.

stateMultiset ∷ Multiset (State n)

States seen so far, with multiplicity.

depth ∷ Integer

The number of edges taken so far.

word ∷ Path n e → [Symbol e]

The reversal of the labels of the Path.

initialsPaths ∷ (Ord e, Ord n) ⇒ FSA n e → Set (Path n e)

Initial open list for traversal from initial states.

initialsNDPath ∷ (Ord e, Ord n) ⇒ FSA n e → Path (Set n) e

rejectingPaths ∷ (Ord e, Ord n) ⇒ FSA n e → Integer → Set (Path n e)

All paths of length less than or equal to a given bound that do not end in an accepting state.

acyclicPaths ∷ (Ord e, Ord n) ⇒ FSA n e → Set (Path n e)

All paths from initialsPaths that do not contain cycles.

extensions ∷ (Ord e, Ord n) ⇒ FSA n e → Path n e → Set (Path n e)

Paths extending a given path by a single edge.

boundedCycleExtensions ∷ (Ord e, Ord n) ⇒ FSA n e → Integer → Path n e → Set (Path n e)

Extensions other than back-edges to a state that has been visited more than a given number of times.

nondeterministicAcyclicExtensions ∷ (Ord e, Ord n) ⇒ FSA n e → Path (Set n) e → Set (Path (Set n) e)